feat: Calculate and validate checksums in outbox storage
